pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "airpilot-blackbox"
version = "0.1.0"
description = "AirPilot Blackbox - Analyze Claude AI usage logs and calculate costs"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = ["airpilot", "blackbox", "claude", "ai", "usage", "analytics", "cost-tracking"]
authors = [
    { name = "shaneholloman" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]

dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "python-dateutil>=2.8.0",
    "requests>=2.31.0",
]

[project.urls]
Homepage = "https://github.com/shaneholloman/airpilot-blackbox"
Repository = "https://github.com/shaneholloman/airpilot-blackbox"
Issues = "https://github.com/shaneholloman/airpilot-blackbox/issues"

[project.scripts]
blackbox = "airpilot_blackbox.cli:main"
airpilot-blackbox = "airpilot_blackbox.cli:main"

[tool.uv]
dev-dependencies = [
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "types-requests>=2.31.0",
    "types-python-dateutil>=2.8.0"
]

[tool.ruff]
target-version = "py38"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
